# cms-node/public/swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "cms-Node接口",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "apiteam@swagger.io"
		}
	},
	"schemes": ["https", "http"],
	"paths": {
    "/node-api/atlas-cms/addPageJson": {
      "post": {
        "tags": ["cms后台接口"],
        "summary": "新增页面json",
        "description": "",
				"operationId": "addPageJson",
        "produces": ["application/json"],
        "parameters": [
          {"name":"Authorization","in":"header","description":"token","required":false,"type":"string"},
          {"in":"body","name":"req","description":"req","required":true, "schema": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "页面名称"
              },
              "color": {
                "type": "string",
                "description": "标题颜色"
              },
              "bgColor": {
                "type": "string",
                "description": "背景色"
              },
              "cover": {
                "type": "string",
                "description": "封面链接"
              },
              "componentList": {
                "type":"array",
                "items": {
                }

              }
            }
          }}
        ],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					}
				}



      }
    },
    "/node-api/atlas-cms/updateCmsJson": {
      "post": {
        "tags": ["cms后台接口"],
        "summary": "编辑页面json",
        "description": "",
				"operationId": "addPageJson",
        "produces": ["application/json"],
        "parameters": [
          {"name":"Authorization","in":"header","description":"token","required":false,"type":"string"},
          {"in":"body","name":"req","description":"req","required":true, "schema": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "id"
              },
              "name": {
                "type": "string",
                "description": "页面名称"
              },
              "color": {
                "type": "string",
                "description": "标题颜色"
              },
              "bgColor": {
                "type": "string",
                "description": "背景色"
              },
              "cover": {
                "type": "string",
                "description": "封面链接"
              },
              "componentList": {
                "type":"array",
                "items": {
                }

              }
            }
          }}
        ],
        "responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					}
				}
      }
    },
    "/node-api/atlas-cms/getPageJson": {
      "get": {
        "tags": ["cms后台接口"],
        "summary": "cms获取当前页数据",
        "description": "",
				"operationId": "getPageJson",
        "produces": ["application/json"],
        "parameters": [
          {"name":"Authorization","in":"header","description":"token","required":false,"type":"string"},
          {"in":"body","name":"req","description":"req","required":true,"schema": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "页面id"
              }
            }
          }}
        ],
        "responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/getPageJsonApiResponse"
						}
					}
				}
      }
    },
    "/node-api/atlas-cms/getPageList": {
      "get": {
        "tags": ["cms后台接口"],
        "summary": "页面列表",
        "description": "",
				"operationId": "getPageList",
        "produces": ["application/json"],
        "parameters": [
          {"name":"Authorization","in":"header","description":"token","required":false,"type":"string"},
          {"in":"body","name":"req","description":"req","required":true,"schema": {
            "type": "object",
            "properties": {
              "pageNum": {
                "type": "integer",
                "description": "起始页数"
              },
              "pageSize": {
                "type": "integer",
                "description": "每页数量"
              }
            }
          }}
        ],
        "responses": {
          "200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/getPageListApiResponse"
						}
					}
        }
      }
    },
    "/node-api/atlas-cms/deletePage": {
      "post": {
        "tags": ["cms后台接口"],
        "summary": "页面删除",
        "description": "",
				"operationId": "deletePage",
        "produces": ["application/json"],
        "parameters": [
          {"name":"Authorization","in":"header","description":"token","required":false,"type":"string"},
          {"in":"body","name":"req","description":"req","required":true,"schema": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "页面id"
              }
            }
          }}
        ],
        "responses": {
          "200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					}
        }
      }
    },
    "/node-api/atlas-cms/getH5PageJson": {
      "get": {
        "tags": ["c端接口"],
				"summary": "组件类目查询",
				"description": "",
				"operationId": "getH5PageJson",
				"produces": ["application/json"],
				"parameters": [
          {"name":"Authorization","in":"header","description":"token","required":false,"type":"string"},
          {"in":"body","name":"req","description":"req","required":true,"schema": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "页面id"
              }
            }
          }}
        ],
        "responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/getH5PageJsonApiResponse"
						}
					}
				}
      }
    },
    "/node-api/atlas-cms/updatePageStatus": {
      "post": {
        "tags": ["cms后台接口"],
        "summary": "更新页面上架状态",
        "description": "",
        "operationId": "updatePageStatus",
        "produces": ["application/json"],
        "parameters": [
          {"name":"Authorization","in":"header","description":"token","required":false,"type":"string"},
          {"in":"body","name":"req","description":"req","required":true,"schema": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "页面id"
              },
              "isAbled": {
                "type": "integer",
                "description": "是否启用 0未启用；1启用"
              }
            }
          }}
        ],
        "responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					}
				}
      }
    }
	},
	"externalDocs": {
		"description": "Find out more about Swagger",
		"url": "http://swagger.io"
	},
	"definitions": {
		"ApiResponse": {
			"type": "object",
			"properties": {
				"code": {
					"type": "integer"
				},
				"data": {
					"type": "string"
				},
				"message": {
					"type": "string"
				}
			}
		},
    "getComponentCategoryApiResponse": {
      "type": "object",
      "properties": {
				"code": {
					"type": "integer"
				},
				"data": {
					"type":"array",
          "items": {
            "properties": {
              "id": {
                "type": "integer",
                "description": "类目id"
              },
              "name": {
                "type": "string",
                "description": "类目名称"
              },
              "component_definitions": {
                "type": "array",
                "items": {
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "组件id"
                    },
                    "name": {
                      "type": "string",
                      "description": "组件名称"
                    },
                    "component_key": {
                      "type": "string",
                      "description": "组件标示"
                    }
                  }
                }
              }
            }
          }

				},
				"message": {
					"type": "string"
				}
			}
    },
    "getH5PageJsonApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {
          "type":"array",
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "页面名称"
              },
              "color": {
                "type": "string",
                "description": "标题颜色"
              },
              "bgColor": {
                "type": "string",
                "description": "背景色"
              },
              "cover": {
                "type": "string",
                "description": "封面链接"
              },
              "componentList": {
                "type":"array",
                "items": {
                }

              }
            }
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getPageListApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {
          "type":"object",
          "properties": {
            "list": {
              "type": "array",
              "items": {
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "id"
                  },
                  "update_time": {
                    "type": "string",
                    "description": "更新时间"
                  },
                  "name": {
                    "type": "string",
                    "description": "页面名称"
                  },
                  "page_type": {
                    "type": "integer",
                    "description": "页面类型 0标准活动；1首页"
                  },
                  "is_abled": {
                    "type": "integer",
                    "description": "是否启用 0未启用；1启用"
                  }
                }
              }
            },
            "pageNum": {
              "type": "integer"
            },
            "pageSize": {
              "type": "integer"
            },
            "total": {
              "type": "integer"
            }
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getPageJsonApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {
          "type":"object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "id"
            },
            "update_time": {
              "type": "string",
              "description": "更新时间"
            },
            "name": {
              "type": "string",
              "description": "页面名称"
            },
            "page_type": {
              "type": "integer",
              "description": "页面类型 0标准活动；1首页"
            },
            "is_abled": {
              "type": "integer",
              "description": "是否启用 0未启用；1启用"
            }
          }
        },
        "message": {
          "type": "string"
        }
      }
    }
	}
}